Maker to Server; consolidate Indexer
