+++ tools/libraries/index.ts
@@ -1,4 +1,4 @@
 export { SwapERC20 } from './src/SwapERC20'
 export { Wrapper } from './src/Wrapper'
 export { Registry } from './src/Registry'
-export { Maker } from './src/Maker'
+export { Server } from './src/Server'
