+++ tools/libraries/package.json
@@ -24,7 +24,7 @@
     "pretty:check": "prettier --check \"./**/*.ts\" \"./**/*.js\"",
     "pretty:fix": "prettier --write \"./**/*.ts\" \"./**/*.js\"",
     "test": "yarn test:unit",
-    "test:unit": "yarn compile && yarn mocha -r ts-node/register test/Maker.test.ts test/Indexer.test.ts"
+    "test:unit": "yarn compile && yarn mocha -r ts-node/register test/Server.test.ts"
   },
   "dependencies": {
     "@airswap/constants": "^4.0.0",
